moved fortran restart from state to initialization
