+++ util/pace/util/initialization/__init__.py
@@ -1,6 +1,2 @@
 from .allocator import QuantityFactory
-from .fortran_restart_to_pace_variable_names import (
-    extra_restart_properties,
-    fortran_restart_to_pace_dict,
-)
 from .sizer import GridSizer, SubtileGridSizer
